requirements.txt
argparse
pytest
psycopg2
cryptography
pycryptodome